refactor: Split variable file to function constants
variable.js -> helper/utils/index.js, constants/regex/index.js
