+++ js/constants/regex/index.js
@@ -1,0 +1,3 @@
+export const emailValidate = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+
+export const passwordPattern = /^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{8,}$/;
